+++ src/constants/MoveHistory/EventTemplates/UpdateMTOShipment/updateMTOShipmentDeprecatePaymentRequest.test.js
@@ -1,6 +1,6 @@
 import o from 'constants/MoveHistory/UIDisplay/Operations';
 import getTemplate from 'constants/MoveHistory/TemplateManager';
-import e from 'constants/MoveHistory/EventTemplates/updateMTOShipmentDeprecatePaymentRequest';
+import e from 'constants/MoveHistory/EventTemplates/UpdateMTOShipment/updateMTOShipmentDeprecatePaymentRequest';
 
 describe('when given a deprecated payment request history record', () => {
   const item = {
