+++ src/constants/MoveHistory/EventTemplates/index.js
@@ -13,8 +13,8 @@ export { default as createEntitlements } from './CreateOrders/createEntitlements
 export { default as createMoves } from './CreateOrders/createMoves';
 export { default as createPaymentRequest } from './CreatePaymentRequest/createPaymentRequest';
 export { default as createPaymentRequestReweighUpdate } from './createPaymentRequestReweighUpdate';
-export { default as createPaymentRequestShipmentUpdate } from './createPaymentRequestShipmentUpdate';
-export { default as createReweighRequestWeightUpdate } from './createReweighRequestWeightUpdate';
+export { default as createPaymentRequestShipmentUpdate } from './UpdateMTOShipment/createPaymentRequestShipmentUpdate';
+export { default as createReweighRequestWeightUpdate } from './UpdateMTOShipment/createReweighRequestWeightUpdate';
 export { default as createStandardServiceItem } from './createStandardServiceItem';
 export { default as deleteShipment } from './DeleteShipment/deleteShipment';
 export { default as proofOfServiceDocUpload } from './CreateUpload/proofOfServiceDocUpload';
@@ -31,8 +31,8 @@ export { default as updateMoveTaskOrder } from './updateMoveTaskOrder';
 export { default as updateMoveTaskOrderStatus } from './UpdateMoveTaskOrderStatus/updateMoveTaskOrderStatus';
 export { default as updateMTOServiceItem } from './UpdateMTOServiceItem/updateMTOServiceItem';
 export { default as updateMTOShipment } from './UpdateMTOShipment/updateMTOShipment';
-export { default as updateMTOShipmentAgent } from './updateMTOShipmentAgent';
-export { default as updateMTOShipmentDeprecatePaymentRequest } from './updateMTOShipmentDeprecatePaymentRequest';
+export { default as updateMTOShipmentAgent } from './UpdateMTOShipment/updateMTOShipmentAgent';
+export { default as updateMTOShipmentDeprecatePaymentRequest } from './UpdateMTOShipment/updateMTOShipmentDeprecatePaymentRequest';
 export { default as reweighPaymentRequest } from './reweighPaymentRequest';
 export { default as updateOrder } from './UpdateOrders/updateOrder';
 export { default as updatePaymentRequest } from './updatePaymentRequest';
@@ -44,7 +44,7 @@ export { default as updateMTOReviewedBillableWeightsAt } from './UpdateMTOReview
 export { default as uploadAmendedOrders } from './UploadAmendedOrders/uploadAmendedOrders';
 export { default as updateMTOShipmentStatus } from './UpdateMTOShipmentStatus/updateMTOShipmentStatus';
 export { default as updateMTOServiceItemMoveStatus } from './CreateMTOServiceItem/createMTOServiceItemUpdateMoveStatus';
-export { default as updateMoveEstimatedExcessWeight } from './updateMTOShipmentPrimeEstimatedExcessWeight';
+export { default as updateMoveEstimatedExcessWeight } from './UpdateMTOShipment/updateMTOShipmentPrimeEstimatedExcessWeight';
 export { default as updateMTOStatusServiceCounselingCompleted } from './updateMTOStatusServiceCounselingCompleted';
 export { default as updateAllowanceByCounselor } from './UpdateAllowances/updateAllowanceByCounselor';
 export { default as updateServiceMemberByCounselor } from './UpdateServiceMember/updateServiceMemberByCounselor';
